Fix issues with SqlTransport_8 sample. Remove InputLoopService
