+++ samples/sqltransport/native-integration/SqlTransport_8/Receiver/Program.cs
@@ -4,25 +4,22 @@ using System.Linq;
 using System.Text;
 using System.Threading.Tasks;
 using Microsoft.Data.SqlClient;
-using Microsoft.Extensions.DependencyInjection;
 using Microsoft.Extensions.Hosting;
 using Newtonsoft.Json;
 using Newtonsoft.Json.Serialization;
 using NServiceBus;
-using Receiver;
 
 // for SqlExpress use Data Source=.\SqlExpress;Initial Catalog=NsbSamplesSqlNativeIntegration;Integrated Security=True;Max Pool Size=100;Encrypt=false
-var connectionString = @"Server=localhost,1433;Initial Catalog=NsbSamplesSqlNativeIntegration;User Id=SA;Password=yourStrong(!)Password;Max Pool Size=100;Encrypt=false";
+//var connectionString = @"Server=localhost,1433;Initial Catalog=NsbSamplesSqlNativeIntegration;User Id=SA;Password=yourStrong(!)Password;Max Pool Size=100;Encrypt=false";
+// for LocalDB
+var connectionString = @"Data Source=(localdb)\MSSQLLocalDB;Initial Catalog=NsbSamplesSqlNativeIntegration;Integrated Security=True;Connect Timeout=30;Encrypt=False;Max Pool Size=100;Trust Server Certificate=False;Application Intent=ReadWrite;Multi Subnet Failover=False";
 
 Console.Title = "NativeIntegration";
 var builder = Host.CreateApplicationBuilder(args);
 
-builder.Services.AddSingleton(provider => new InputLoopService(connectionString));
-builder.Services.AddHostedService(sp => sp.GetRequiredService<InputLoopService>());
-
 #region EndpointConfiguration
-var endpointConfiguration = new EndpointConfiguration("Samples.SqlServer.NativeIntegration");
 
+var endpointConfiguration = new EndpointConfiguration("Samples.SqlServer.NativeIntegration");
 endpointConfiguration.UseTransport(new SqlServerTransport(connectionString)
 {
     TransportTransactionMode = TransportTransactionMode.SendsAtomicWithReceive
@@ -33,6 +30,7 @@ endpointConfiguration.UseSerialization<NewtonsoftJsonSerializer>()
         TypeNameHandling = TypeNameHandling.Auto,
         SerializationBinder = new SkipAssemblyNameForMessageTypesBinder([typeof(PlaceOrder), typeof(LegacyOrderDetected)])
     });
+
 #endregion
 
 endpointConfiguration.EnableInstallers();
@@ -42,11 +40,61 @@ await SqlHelper.EnsureDatabaseExists(connectionString);
 Console.WriteLine("Press enter to send a message");
 Console.WriteLine("Press any key to exit");
 
-
 builder.UseNServiceBus(endpointConfiguration);
 
-await builder.Build().RunAsync();
+var host = builder.Build();
+await host.StartAsync();
 
+while (true)
+{
+    var key = Console.ReadKey();
+    Console.WriteLine();
+
+    if (key.Key != ConsoleKey.Enter)
+    {
+        break;
+    }
+
+    await PlaceOrder(connectionString);
+}
+
+await host.StopAsync();
+
+static async Task PlaceOrder(string connectionString)
+{
+    #region MessagePayload
+
+    var message = @"{
+                           $type: 'PlaceOrder, Receiver',
+                           OrderId: 'Order from ADO.net sender'
+                        }";
+
+    #endregion
+
+    #region SendingUsingAdoNet
+
+    var insertSql = @"insert into [Samples.SqlServer.NativeIntegration]
+                                      (Id, Recoverable, Headers, Body)
+                               values (@Id, @Recoverable, @Headers, @Body)";
+    using (var connection = new SqlConnection(connectionString))
+    {
+        await connection.OpenAsync();
+
+        using (var command = new SqlCommand(insertSql, connection))
+        {
+            var parameters = command.Parameters;
+            parameters.Add("Id", SqlDbType.UniqueIdentifier).Value = Guid.NewGuid();
+            parameters.Add("Headers", SqlDbType.NVarChar).Value = "";
+            var body = Encoding.UTF8.GetBytes(message);
+            parameters.Add("Body", SqlDbType.VarBinary).Value = body;
+            parameters.Add("Recoverable", SqlDbType.Bit).Value = true;
+
+            await command.ExecuteNonQueryAsync();
+        }
+    }
+
+    #endregion
+}
 
 class SkipAssemblyNameForMessageTypesBinder(Type[] messageTypes) : ISerializationBinder
 {
